+++ project-loader/src/main/kotlin/de/itemis/mps/gradle/project/loader/Macro.kt
@@ -1,0 +1,6 @@
+package de.itemis.mps.gradle.project.loader
+
+public data class Macro(
+        val name: String,
+        val value: String
+)