+++ project-loader/src/main/kotlin/de/itemis/mps/gradle/project/loader/Plugin.kt
@@ -1,0 +1,6 @@
+package de.itemis.mps.gradle.project.loader
+
+public data class Plugin(
+        val id: String,
+        val path: String
+)